requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==25.1.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
datasets==2.21.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy==2.6.10
exceptiongroup==1.2.2
fastapi==0.115.11
fastembed==0.6.0
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.6.1
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
json_repair==0.39.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.63.3
loguru==0.7.3
magicattr==0.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
numpy==2.2.3
onnxruntime==1.21.0
openai==1.65.5
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
propcache==0.3.0
protobuf==5.29.3
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
psycopg2
PyYAML==6.0.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
qdrant-client==1.13.3
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resend==2.6.0
rpds-py==0.23.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
stripe==11.6.0
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
win32_setctime==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
